Change convert() signature; now returns invocant.
Added content() to return the converted content from the previous call
to convert() (i.e. what convert() used to return).

Added modified() to return a true value if the previous call to
convert() actually modified the file, and false if not. The actual
return is the number of modifications, but this is an implementation
detail and is intentionally undocumented.